refactor - extract azure transmission logic to a mixin
